requirements.txt
numpy==2.3.5
scipy==1.16.3
matplotlib==3.10.8
pillow==12.0.0
plotly==6.5.0
streamlit==1.52.1
ipykernel==7.1.0
